Re-wrote history to remove bogus commits (sorry), and added code to try to avoid them.
Since I added the -A flag to the "git add" call, an error during
retrieval which causes a chunk of the law to not get downloaded
then caused a commit with all of the un-downloaded law removed.
The following day, of course, a successful retrieval would then
result in a commit with the law re-added.  This change attempts
to notice when retrieval fails and to skip doing the git commit
in that case.